Upgrade dependencies to their latest versions
Kotlin 1.9.10, Java 20, Compose 1.5.1, Xcode 15, iOS 17 and more

+++ Shared/src/macosMain/kotlin/com/conradkramer/wallet/ViewServiceServer.kt
@@ -24,7 +24,7 @@ class ViewServiceServer internal constructor(
 ) {
     private val queue = dispatch_queue_create("$machService.listener", null)
     private val listener =
-        xpc_connection_create_mach_service(machService, queue, XPC_CONNECTION_MACH_SERVICE_LISTENER)
+        xpc_connection_create_mach_service(machService, queue, XPC_CONNECTION_MACH_SERVICE_LISTENER.toULong())
 
     private val connections: MutableMap<Session, xpc_connection_t> = mutableMapOf()
